requirements.txt
# for ml related requirments see ml.requirments.txt
beautifulsoup4==4.12.3
boto3==1.35.4
boto3-stubs
botocore==1.35.4
botocore-stubs
brotlipy==0.7.0
click==8.1.7 # CLI construction
databricks-sql-connector==3.3.0
dataset==1.6.2
deflate==0.7.0
mypy==1.11.1  # static type checking for better quality code
pandas==2.2.2
pandas-stubs~=2.2.2
pyarrow==17.0.0
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.31.0
types-requests~=2.31
types-beautifulsoup4
rich==13.7.1  # colored logging
